# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotel Guest Assistant API",
    "description": "Multi-tenant hotel guest assistant. Guest API under /api/v1/hotels/{hotel_id}; legacy /api/* endpoints are deprecated.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/hotels/{hotel_id}": {
      "get": {
        "tags": [
          "guest v1"
        ],
        "summary": "Hotel Profile",
        "description": "Public, non-sensitive hotel profile for the guest UI (branding, languages, form limits).",
        "operationId": "hotel_profile_api_v1_hotels__hotel_id__get",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hotel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        }
      }
    },
    "/api/v1/hotels/{hotel_id}/conversations": {
      "post": {
        "tags": [
          "guest v1"
        ],
        "summary": "Create Conversation",
        "operationId": "create_conversation_api_v1_hotels__hotel_id__conversations_post",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hotel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationCreated"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        }
      }
    },
    "/api/v1/hotels/{hotel_id}/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "guest v1"
        ],
        "summary": "Get Conversation",
        "operationId": "get_conversation_api_v1_hotels__hotel_id__conversations__conversation_id__get",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hotel Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        }
      },
      "delete": {
        "tags": [
          "guest v1"
        ],
        "summary": "Delete Conversation",
        "description": "Guest-initiated deletion (data minimisation).",
        "operationId": "delete_conversation_api_v1_hotels__hotel_id__conversations__conversation_id__delete",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hotel Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        }
      }
    },
    "/api/v1/hotels/{hotel_id}/conversations/{conversation_id}/messages": {
      "post": {
        "tags": [
          "guest v1"
        ],
        "summary": "Post Message",
        "operationId": "post_message_api_v1_hotels__hotel_id__conversations__conversation_id__messages_post",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hotel Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationTurnResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        }
      }
    },
    "/api/v1/hotels/{hotel_id}/conversations/{conversation_id}/availability": {
      "post": {
        "tags": [
          "guest v1"
        ],
        "summary": "Conversation Availability",
        "description": "Booking-form submission: deterministic search, recorded in the conversation's context.",
        "operationId": "conversation_availability_api_v1_hotels__hotel_id__conversations__conversation_id__availability_post",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hotel Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/hotels/{hotel_id}/availability": {
      "post": {
        "tags": [
          "guest v1"
        ],
        "summary": "Availability",
        "description": "Stateless availability search (e.g. for a booking-engine widget).",
        "operationId": "availability_api_v1_hotels__hotel_id__availability_post",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hotel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
    "/api/availability": {
      "post": {
        "tags": [
          "legacy (deprecated)"
        ],
        "summary": "Availability",
        "operationId": "availability_api_availability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResult"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    }
  },
  "components": {
    "schemas": {
      "AvailabilityRequest": {
        "properties": {
          "check_in": {
            "type": "string",
            "format": "date",
            "title": "Check In"
          },
          "check_out": {
            "type": "string",
            "format": "date",
            "title": "Check Out"
          },
          "adults": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Adults"
          },
          "children": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Children",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "check_in",
          "check_out",
          "adults"
        ],
        "title": "AvailabilityRequest"
      },
      "AvailabilityResult": {
        "properties": {
          "check_in": {
            "type": "string",
            "format": "date",
            "title": "Check In"
          },
          "check_out": {
            "type": "string",
            "format": "date",
            "title": "Check Out"
          },
          "nights": {
            "type": "integer",
            "title": "Nights"
          },
          "adults": {
            "type": "integer",
            "title": "Adults"
          },
          "children": {
            "type": "integer",
            "title": "Children"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/RoomOffer"
            },
            "type": "array",
            "title": "Rooms"
          },
          "sold_out_room_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sold Out Room Names",
            "default": []
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "season_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Label"
          }
        },
        "type": "object",
        "required": [
          "check_in",
          "check_out",
          "nights",
          "adults",
          "children",
          "available",
          "rooms",
          "message"
        ],
        "title": "AvailabilityResult"
      },
      "BookingContext": {
        "properties": {
          "check_in": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check In"
          },
          "check_out": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Out"
          },
          "adults": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Adults"
          },
          "children": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Children"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BookingContext",
        "description": "Last booking details the guest entered, so follow-ups like 'what about 3 adults?' work."
      },
...
      "ChatReply": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "answer",
              "clarification",
              "fallback",
              "availability",
              "collect_booking_details"
            ],
            "title": "Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions",
            "default": []
          },
          "availability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AvailabilityResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "booking_prefill": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingPrefill"
              },
              {
                "type": "null"
              }
            ]
          },
          "form_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Error"
          }
        },
        "type": "object",
        "required": [
          "type",
          "text"
        ],
        "title": "ChatReply"
      },
...
      "ConversationCreated": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "hotel_id": {
            "type": "string",
            "title": "Hotel Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "hotel_id",
          "channel",
          "locale",
          "expires_at"
        ],
        "title": "ConversationCreated"
      },
      "ConversationTurnResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "ai",
              "offline"
            ],
            "title": "Mode"
          },
          "reply": {
            "$ref": "#/components/schemas/ChatReply"
          },
          "notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "conversation_id",
          "mode",
          "reply",
          "meta"
        ],
        "title": "ConversationTurnResponse"
      },
      "ConversationView": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "hotel_id": {
            "type": "string",
            "title": "Hotel Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "active_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Intent"
          },
          "availability_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageView"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "hotel_id",
          "channel",
          "locale",
          "created_at",
          "updated_at",
          "expires_at",
          "active_intent",
          "availability_context",
          "messages"
        ],
        "title": "ConversationView"
      },
      "CreateConversationRequest": {
        "properties": {
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z]{2}(-[A-Z]{2})?$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateConversationRequest"
      },
...
      "ErrorResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "MessageView": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "reply_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "created_at"
        ],
        "title": "MessageView"
      },
      "PostMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z]{2}(-[A-Z]{2})?$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "PostMessageRequest"
      },
...
      "V1ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "V1ErrorBody"
      },
      "V1ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/V1ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "V1ErrorResponse"
      },
...
    }
  }
}